【BUG】 修复内存交易记录和个人交易记录解析失败错误 【功能】 添加内存交易记录跳转个人交易记录逻辑

+++ app/src/main/java/io/goldstone/blockchain/module/home/quotation/tradermemory/eosmemorytransactionhistorylist/model/EOSMemoryTransactionHistoryListTable.kt
@@ -12,7 +12,7 @@ import com.google.gson.annotations.SerializedName
 data class EOSMemoryTransactionHistoryListTable(
 	val account: String,
 	val price: Double,
-	val quantity: Int,
+	val quantity: Double,
 	val time: Long,
 	val txId: String,
 	val type: Int
@@ -21,7 +21,7 @@ data class EOSMemoryTransactionHistoryListTable(
 	constructor() : this(
 		"",
 		0.0,
-		0,
+		0.0,
 		0,
 		"",
 		0
@@ -49,7 +49,7 @@ data class EOSMemoryTransactionHistoryListModel(
 		@SerializedName("price")
 		val price: Double,
 		@SerializedName("quantity")
-		val quantity: Int,
+		val quantity: Double,
 		@SerializedName("time")
 		val time: Long,
 		@SerializedName("tx_id")
